+++ CHANGES.md
@@ -139,6 +139,16 @@ To be released.
 [#3044]: https://github.com/planetarium/libplanet/pull/3044
 
 
+Version 0.53.3
+--------------
+
+Released on April 3, 2023.
+
+ -  Adjusted level of the consensus related logs.  [[#3046]]
+
+[#3046]: https://github.com/planetarium/libplanet/pull/3046
+
+
 Version 0.53.2
 --------------
 
@@ -146,11 +156,11 @@ Released on March 31, 2023.
 
  -  Ported changes from [Libplanet 0.50.7] release.  [[#3022]]
  -  `BlockChain<T>.Create()` is now get `IBlockChainState` and
-    `ActionEvaluator` as parameters.  [[#3029]]
+    `ActionEvaluator` as parameters.  [[#3028]]
 
 [Libplanet 0.50.7]: https://www.nuget.org/packages/Libplanet/0.50.7
 [#3022]: https://github.com/planetarium/libplanet/pull/3022
-[#3029]: https://github.com/planetarium/libplanet/pull/3029
+[#3028]: https://github.com/planetarium/libplanet/pull/3028
 
 
 Version 0.53.1
@@ -569,10 +579,17 @@ Version 0.50.7
 
 Released on March 29, 2023.
 
+<<<<<<< HEAD
  -  Optimized internal byte array to hex string conversion in several places.
     [[#2989], [#3007]]
  -  `IValidatorSupportStateDelta` is now public [[#3017]]
  -  Optimized `BlockChainState.GetStates()` by caching results.  [[#3018]]
+=======
+-  Optimized internal byte array to hex string conversion in several places.
+   [[#2989], [#3007]]
+-  `IValidatorSupportStateDelta` is now public [[#3017]]
+-  Optimized `BlockChainState.GetStates()` by caching results.  [[#3018]]
+>>>>>>> 0.53.3
 
 [#2989]: https://github.com/planetarium/libplanet/issues/2989
 [#3007]: https://github.com/planetarium/libplanet/pull/3007
@@ -585,8 +602,13 @@ Version 0.50.6
 
 Released on March 22, 2023.
 
+<<<<<<< HEAD
  -  Remove `SetValidator` system action.  [[#2972]]
  -  `ValidatorStateExtensions` is now public.  [[#2972]]
+=======
+-  Remove `SetValidator` system action.  [[#2972]]
+-  `ValidatorStateExtensions` is now public.  [[#2972]]
+>>>>>>> 0.53.3
 
 [#2972]: https://github.com/planetarium/libplanet/pull/2972
 
@@ -598,8 +620,13 @@ Released on March 17, 2023.
 
 ### Bug fixes
 
+<<<<<<< HEAD
  -  (Libplanet.Net) Fixed a bug where `DealerSocket` cannot connect to peers
     because of IPv4 return of `ResolveNetMQAddressAsync()`.  [[#2956]]
+=======
+-  (Libplanet.Net) Fixed a bug where `DealerSocket` cannot connect to peers
+   because of IPv4 return of `ResolveNetMQAddressAsync()`.  [[#2956]]
+>>>>>>> 0.53.3
 
 [#2956]: https://github.com/planetarium/libplanet/pull/2956